update cert to DB

+++ ec-contract-srv/src/main/java/com/ec/contract/repository/CertificateMappingRepository.java
@@ -6,6 +6,6 @@ import com.ec.contract.model.entity.keystoreEntity.CertificateCustomer;
 import com.ec.contract.model.entity.keystoreEntity.CertificateMapping;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CertificateMappingRepository extends JpaRepository<CertificateMapping, Long> {
+public interface CertificateMappingRepository extends JpaRepository<CertificateMapping, Integer> {
     void deleteCertificateMappingByCertificateAndCertificateCustomer(Certificate certificate, CertificateCustomer customer);
 }
